--- config_example.ini
@@ -1,5 +1,0 @@
-[twitter_credentials]
-consumer_key = <>
-consumer_secret = <>
-access_token = <>
-access_token_secret = <>
